アカウント管理画面UI修正

+++ src/main/java/com/github/rk_aiz/teamsurvey/domain/service/impl/UserGroupServiceImpl.java
@@ -1,13 +1,16 @@
 package com.github.rk_aiz.teamsurvey.domain.service.impl;
 
 import java.util.List;
+import java.util.Objects;
 
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.github.rk_aiz.teamsurvey.domain.model.LoginUser;
+import com.github.rk_aiz.teamsurvey.domain.exception.ServiceRuleException;
 import com.github.rk_aiz.teamsurvey.domain.model.UserGroup;
 import com.github.rk_aiz.teamsurvey.domain.service.UserGroupService;
+import com.github.rk_aiz.teamsurvey.domain.type.Authority;
 import com.github.rk_aiz.teamsurvey.infrastructure.repository.UserGroupRepository;
 
 import lombok.RequiredArgsConstructor;
@@ -16,10 +19,10 @@ import lombok.RequiredArgsConstructor;
 @Transactional
 @RequiredArgsConstructor
 public class UserGroupServiceImpl implements UserGroupService {
-    
+
     /** DI */
     private final UserGroupRepository userGroupRepository;
-    
+
     @Override
     public List<UserGroup> findAll() {
         return userGroupRepository.findAll();
@@ -27,15 +30,37 @@ public class UserGroupServiceImpl implements UserGroupService {
 
     @Override
     public boolean save(UserGroup userGroup) {
-        if (userGroup.getGroupId() == null) {
-            return userGroupRepository.add(userGroup);
+
+        if (userGroup.getId() == null) {
+            // 新規登録時のチェック
+            boolean isDuplicate = this.userGroupRepository
+                    .existsByGroupName(userGroup.getGroupName());
+
+            if (isDuplicate) {
+                throw new DuplicateKeyException("指定されたグループ名は既に使用されています。");
+            }
         } else {
-            return userGroupRepository.set(userGroup);
+            UserGroup currentDbGroup = this.userGroupRepository
+                    .findById(userGroup.getId());
+
+            // システムグループの場合は名称の変更のみ許可
+            if (currentDbGroup.isSystemGroup() &&
+                    currentDbGroup.getAuthority() != userGroup.getAuthority()) {
+                throw new ServiceRuleException("システムグループは権限の変更はできません");
+            }
         }
+        return this.userGroupRepository.save(userGroup);
     }
 
     @Override
     public boolean delete(Integer groupId) {
+        UserGroup targetGroup = this.userGroupRepository.findById(groupId);
+
+        // グループが存在しない、またはシステムグループの場合は削除不可
+        if (targetGroup == null || targetGroup.isSystemGroup()) {
+            return false;
+        }
+
         return userGroupRepository.remove(groupId);
     }
 }
